Introduce CHERI feature flags
Probably not intended for standardization, but intended to ease our
roll-out of features through multiple cores with the side benefit of
potentially enabling science (TM) if these bits are toggleable on real
cores.

Right now, just define scisa, as the only features we're tracking are
S-mode.  I'm not sure if we will ever need or want M-mode or U-mode
feature flags, but if we do, the current set of bits is not really
relevant to those modes.